funguje updating cen na exchange, ale blbne nám opět propisování account classy do dalších formů, v exchange není vidět jako co se člověk přihlásil

+++ Client-side form/Client-side form/_Account.cs
@@ -11,7 +11,7 @@ namespace Client_side_form
     {
         public string userName { get; set; }
         public string password { get; set; }
-        public int balance { get; set; }
+        public double balance { get; set; }
         public int volume1 { get; set; }
         public int volume2 { get; set; }
         public int volume3 { get; set; }
